+++ packages/teams-js/test/private/videoEx.spec.ts
@@ -4,7 +4,7 @@ import { VideoPerformanceMonitor } from '../../src/internal/videoPerformanceMoni
 import { videoEx } from '../../src/private/videoEx';
 import { app } from '../../src/public/app';
 import { errorNotSupportedOnPlatform, FrameContexts } from '../../src/public/constants';
-import { videoEffects } from '../../src/public/video';
+import { videoEffects } from '../../src/public/videoEffects';
 import { Utils } from '../utils';
 
 /* eslint-disable */
